Fixed test.
This test was broken before, but accidentally was passing.  Default
values for store_true and store_false actions were fixed when
introducing the first simple example.

Those changes revealed a problem in
   test_one_option_missing_argument()

The test clearly requires an option that does not have a default
value.  The original action was 'store_true' which has a default and
thus there is always an option returned by parse().