Compiler workarounds.
pFUnit 4.0 use cases apparently stressed both gfortran 8.3 and ifort
18.0.3 a bit too far.  Issues seem to be around the intersection of
unlimited polymorphism and deferred length strings as well as
intrinsic assignment of deep data structures involving unlimited
polymorphism.

Worth noting that even NAG had issues with some of the attempted workarounds.

One reproducer was submitted to Intel, but not for the numerous other
issues that were encountered.  I was too desperate to get back to a
stable develop branch. (And the even the one reproducer took nearly 6
hours of effort to isolate and boil down.)